Refactor functest storagecluster install into pkg
