+++ pkg/deploy-manager/config.go
@@ -1,0 +1,89 @@
+package deploymanager
+
+import (
+	"fmt"
+	"os"
+
+	ocsv1 "github.com/openshift/ocs-operator/pkg/apis/ocs/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/serializer"
+	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
+)
+
+// InstallNamespace is the namespace ocs is installed into
+const InstallNamespace = "openshift-storage"
+
+// DefaultStorageCluster is the name of the storage cluster the test suite installs
+const DefaultStorageCluster = "test-storagecluster"
+
+// DefaultStorageClassRBD is the name of the ceph rbd storage class the test suite installs
+const DefaultStorageClassRBD = DefaultStorageCluster + "-ceph-rbd"
+
+// MinOSDsCount represents the minimum number of OSDs required for this testsuite to run.
+const MinOSDsCount = 3
+
+func init() {
+	ocsv1.SchemeBuilder.AddToScheme(scheme.Scheme)
+}
+
+// DeployManager is a util tool used by the functional tests
+type DeployManager struct {
+	k8sClient      *kubernetes.Clientset
+	ocsClient      *rest.RESTClient
+	parameterCodec runtime.ParameterCodec
+}
+
+// GetK8sClient is the function used to retrieve the kubernetes client
+func (t *DeployManager) GetK8sClient() *kubernetes.Clientset {
+	return t.k8sClient
+}
+
+// NewDeployManager is the way to create a DeployManager struct
+func NewDeployManager() (*DeployManager, error) {
+	codecs := serializer.NewCodecFactory(scheme.Scheme)
+	parameterCodec := runtime.NewParameterCodec(scheme.Scheme)
+
+	kubeconfig := os.Getenv("KUBECONFIG")
+	if kubeconfig == "" {
+		return nil, fmt.Errorf("No KUBECONFIG environment variable set")
+	}
+
+	// K8s Core api client
+	config, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
+	if err != nil {
+		return nil, err
+	}
+	config.NegotiatedSerializer = serializer.DirectCodecFactory{CodecFactory: scheme.Codecs}
+	config.APIPath = "/apis"
+	config.ContentType = runtime.ContentTypeJSON
+	k8sClient, err := kubernetes.NewForConfig(config)
+	if err != nil {
+		return nil, err
+	}
+
+	// ocs Operator rest client
+	ocsConfig, err := clientcmd.BuildConfigFromFlags("", kubeconfig)
+	if err != nil {
+		return nil, err
+	}
+	ocsConfig.GroupVersion = &ocsv1.SchemeGroupVersion
+	ocsConfig.NegotiatedSerializer = serializer.DirectCodecFactory{CodecFactory: codecs}
+	ocsConfig.APIPath = "/apis"
+	ocsConfig.ContentType = runtime.ContentTypeJSON
+	if ocsConfig.UserAgent == "" {
+		ocsConfig.UserAgent = rest.DefaultKubernetesUserAgent()
+	}
+	ocsClient, err := rest.RESTClientFor(ocsConfig)
+	if err != nil {
+		return nil, err
+	}
+
+	return &DeployManager{
+		k8sClient:      k8sClient,
+		ocsClient:      ocsClient,
+		parameterCodec: parameterCodec,
+	}, nil
+}
